intptr_t: fix void* cast

+++ 30__c__pthreads/pthread__mutex00/arrays.c
@@ -12,6 +12,7 @@
 
 #include <pthread.h>
 #include <unistd.h>
+#include <stdint.h> /* intptr_t */
 
 extern int pthread_setconcurrency(int);
 extern int pthread_getconcurrency();
@@ -35,7 +36,6 @@ int main()
 
   // 5 threads
   pthread_t thr[5];
-  
 
   // init
   puts("init..");
@@ -48,20 +48,20 @@ int main()
     data.int_val[idx] = 0;
     data.double_val[idx] = 0.0;
   }
-  
+
   // set concurrency and create the threads
   printf("the current level is %d set the concurrency level to 4\n", pthread_getconcurrency());
   pthread_setconcurrency(4);
 
   // create threads
   puts("creating threads");
-  for(idx = 0; idx < 5; ++idx){
-    if(0 != pthread_create( (void*) &thr[idx], NULL, add_to_value, (void*) (2*idx))){
+  for(idx = 0; idx < 5; ++idx){ // safe cast: int -> intptr_t -> void*
+    if(0 != pthread_create( (void*) &thr[idx], NULL, add_to_value, (void*) ((intptr_t)(2*idx)))){
       perror("pthread_create failed:");
       exit(EXIT_FAILURE);
     }
   }
-  
+
   // wait till all threads have finished
   puts("joining threads");
   for(idx = 0; idx < 5; ++idx){
@@ -72,7 +72,7 @@ int main()
       exit(EXIT_FAILURE);
     }
   }
-  
+
   // print the results
   printf("Final Values.....\n");
   for(idx = 0; idx < 5; ++idx){
@@ -91,22 +91,22 @@ int main()
 
 
 
-/* 
+/*
    thread implementation
 //*/
 void *add_to_value(void *arg)
 {
-  int inval = (int) arg;
+  // safe cast: void* -> intptr_t -> int
+  int inval = (int) (intptr_t) arg;
 
   printf("\nthread %d: was called..\nlock\taction\n", (inval/2));
 
   /*
     set the mutex on writing into the data structure
-    set a value to both structures and unlock - repeat 
+    set a value to both structures and unlock - repeat
     this procedure 5 times (at least once per entry)
-    
-    observe the order of the working threads on the structure 
-    in the output!
+
+    observe the order of the working threads on the structure in the output!
   //*/
   int idx = 0;
   for(idx = 0; idx < 5; ++idx){
@@ -124,7 +124,7 @@ void *add_to_value(void *arg)
 
     printf("%d -\tthread %d: unlockes!\n", (inval/2), (inval/2));
     pthread_mutex_unlock(&data.data_lock[idx % 5]);
-  }    
+  }
 
   printf("%d -\tthread %d: exits\n", (inval/2), (inval/2));
   pthread_exit((void*) 5);
